+++ parkhangs-webapp/package.json
@@ -7,9 +7,12 @@
     "google-map-react": "^1.1.7",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "^0.9.5"
+    "react-scripts": "^0.9.5",
+    "redux": "^4.0.5"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "react-redux": "^7.2.0"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
