+++ parkhangs-webapp/src/components/DetailModal.css
