+++ parkhangs-webapp/src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import parksReducer from '../features/parks/parksSlice'
+
+export default combineReducers({
+  parks: parksReducer,
+})
